putting pokemons data in a separate file
